fix(pnpr): bound libsql startup connect and share the auth timeout helper
Wrap both Builder::build() paths (remote replica and remote) in
with_auth_timeout(DEFAULT_STARTUP_TIMEOUT, ...) so a stalled or
blackholed libsql endpoint cannot hang pnpr startup indefinitely,
matching the bounded-startup behavior the sqlx backend already has. The
build futures are Box::pin'd so the one-time startup connect does not
enlarge the serve future past the large_futures threshold.

Deduplicate with_auth_timeout: move it into the shared auth module so
both the libsql and sqlx backends use one helper instead of two copies.
Its doc no longer claims libsql writes rely on a statement timeout
(libsql configures none); it states the real invariant that writes must
await directly to avoid an unknown commit state.

Trim the timeout constant doc comments to a contract-focused line each,
dropping the DoS narrative and PR-history references.
